requirements.txt
Django
djangorestframework==3.14.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
django-phonenumber-field
djangorestframework_simplejwt